+++ vite.config.ts
@@ -5,9 +5,12 @@ import dts from 'unplugin-dts/vite';
 export default defineConfig({
   build: {
     lib: {
-      entry: resolve(__dirname, 'lib/index.ts'),
-      name: 'appwrite-typed-db',
-      fileName: format => `appwrite-typed-db.${format}.js`,
+      entry: {
+        appwrite: resolve(__dirname, 'lib/appwrite.ts'),
+        'node-appwrite': resolve(__dirname, 'lib/node-appwrite.ts'),
+      },
+      formats: ['es', 'cjs'],
+      fileName: (format, name) => `${name}.${format}.js`,
     },
   },
   plugins: [dts({ bundleTypes: true })],